fix(gf2-core): fix correctness gaps in generic GF(2^m) over UintExt
- Tighten constructor bound: m < V::BITS (was <=), preventing
  shift-wrapping in element() when m == V::BITS
- Move u64-dependent methods (order, verify_primitive,
  is_irreducible_rabin, prime_factors_of_order_static) to concrete
  impl Gf2mField_ block since they use 1usize/1u64 << m
- Add generic order_v() -> V that works for any backing type
- Change order() return type from usize to u64 to avoid overflow on
  32-bit targets; update BCH call sites accordingly
- Fix all primitivity code to use 1u64 << m instead of 1usize << m,
  and change compute_x_power_value to take u64 exponents
- Restore binary Display format ({:#b}) for Gf2mElement_, adding
  Binary bound to UintExt trait
- Replace TypeId-based SIMD dispatch with compile-time IS_U64 const
  on UintExt trait
- Add UintExt sealed trait in separate uint_ext module with
  implementations for u8, u16, u32, u64, u128
- Add cross-width tests: GF(2^4) over u8/u16/u64/u128, GF(2^7)
  over u8, exhaustive cross-width consistency, m==BITS rejection,
  order_v correctness, binary display format
- Fix doc tests using Gf2mPoly_:: to use Gf2mPoly:: (type alias)

+++ crates/gf2-coding/src/bch/core.rs
@@ -124,7 +124,7 @@ impl BchCode {
     /// ```
     pub fn new(n: usize, k: usize, t: usize, field: Gf2mField) -> Self {
         assert!(n > k, "Codeword length must exceed message length");
-        assert!(n < field.order(), "n must divide 2^m - 1");
+        assert!((n as u64) < field.order(), "n must divide 2^m - 1");
         assert!(t > 0, "Error correction capability must be positive");
 
         // Ensure field has tables (needed for primitive element)
@@ -182,7 +182,7 @@ impl BchCode {
         generator: Gf2mPoly,
     ) -> Self {
         assert!(n > k, "Codeword length must exceed message length");
-        assert!(n < field.order(), "n must divide 2^m - 1");
+        assert!((n as u64) < field.order(), "n must divide 2^m - 1");
         assert!(t > 0, "Error correction capability must be positive");
 
         let designed_distance = 2 * t + 1;
